requirements.txt
yt-dlp==2024.07.09
requests==2.32.3
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
schedule==1.2.1
python-dotenv==1.0.0
flask-wtf==1.2.1
Flask-Migrate==4.0.5
rq==1.16.1
Flask-Moment==1.0.6